Make the `Tensor` class properly abstract.
It should not be possible to directly instantiate this class (only
backend-specific or otherwise subclasses).
Not all subclasses directly store any `data`, they may instead delegate
this to another object that they store (like, in case of `DiscretizedSpaceElement`,
another instance of a `Tensor` subclass).
